Add files via upload
Assessing passive voice sentences. In this file I use a linguistic rule to determine if a sentence is written in a passive or active voice.  
According to  the SEC guidelines passive voice language is harder to comprehend so I calculate the percentage of passive voice for each document : the more the harder.